+++ src/pages/ProductAdd.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AddForm2 from '../components/waste/AddForm2';
+import AddForm from '../components/waste/AddForm';
 import Header from '../components/common/header/Header';
 import axios from 'axios';
 
@@ -7,7 +7,7 @@ const ProductAdd = () => {
   return (
     <div>
       <Header />
-      <AddForm2 />
+      <AddForm />
     </div>
   );
 };
